refactor(chat): extract group-membership + reprocess/regen clusters from the god file
Two more cohesive, contiguous clusters moved into part-file extensions (verbatim,
zero behaviour change) to shrink chat_service.dart:

- chat_service_group_membership.dart — forkToGroupChat, _createGroupMember,
  addCharacterToGroup, removeCharacterFromGroup (group membership lifecycle).
- chat_service_reprocess.dart — manualReprocessNeeds, revertNeedsReprocess,
  regenerateLastMessage (turn replay with the per-speaker realism dance + the
  time-travel realism rollback). regenerateLastMessage alone was ~380 lines.

Neither cluster declares fields or references statics, and none are overridden by
the golden FakeChatService, so the move is safe. chat_service.dart 9,933 → 8,912
lines (−1,021 this pass; ~10,506 → 8,912 across the whole effort).

flutter analyze clean; full unit suite (1348) + parity gate green.
