+++ lib/services/chat_service.dart
@@ -91,6 +91,8 @@ part 'chat/chat_service_evolution.dart';
 part 'chat/chat_service_sillytavern.dart';
 part 'chat/chat_service_group_realism_helpers.dart';
 part 'chat/chat_service_history.dart';
+part 'chat/chat_service_group_membership.dart';
+part 'chat/chat_service_reprocess.dart';
 
 // Internal flag to signal a cancellation request for realism evaluation.
 // This is a file-scope flag to avoid needing to thread state through the
@@ -3278,355 +3280,6 @@ class ChatService extends ChangeNotifier {
     notifyListeners();
   }
 
-  /// Fork the current 1:1 chat into a new group chat, copying all messages.
-  /// The original 1:1 session remains untouched.
-  Future<GroupChat?> forkToGroupChat(
-    List<CharacterCard> additionalCharacters,
-    GroupChatRepository groupRepo, {
-    String? groupName,
-    String? scenario,
-    TurnOrder turnOrder = TurnOrder.roundRobin,
-    Map<String, ({String text, bool creative})> entrances = const {},
-  }) async {
-    if (_isGenerating) return null;
-    if (_activeCharacter == null || _characterRepository == null) return null;
-    if (_messages.isEmpty) return null;
-    // 1:1 → group only. Forking from an existing group would rebuild a group
-    // from just the active speaker (dropping the other members), so refuse it —
-    // use "Add Character to Group" for an existing group instead.
-    if (_activeGroup != null) return null;
-
-    final originalCharId = _getCharacterIdFromCard(_activeCharacter!);
-
-    // Build a default group name
-    final name = groupName?.isNotEmpty == true
-        ? groupName!
-        : [
-            _activeCharacter!.name,
-            ...additionalCharacters.map((c) => c.name),
-          ].join(' & ');
-
-    // Create the group
-    final group = GroupChat(
-      id: 'group_${DateTime.now().millisecondsSinceEpoch}',
-      name: name,
-      // characterIds removed (decoupled). Members handled via group_members + private storage.
-      turnOrder: turnOrder,
-      scenario: scenario ?? '',
-      // v31 columns — new groups start with clean defaults.
-      // baselineRealismState remains '{}' until the caller (or UI) seeds explicit values.
-      baselineRealismState: '{}',
-    );
-    await groupRepo.save(group);
-
-    // Rotation order rule for the new group: original participant(s) first,
-    // then arrivals WITH an entrance (in the order added), then arrivals
-    // WITHOUT an entrance at the end. Member insertion order *is* the
-    // round-robin order (the members table has no explicit sort column), so we
-    // insert in exactly that order.
-    bool hasEntrance(CharacterCard c) =>
-        (entrances[_getCharacterIdFromCard(c)]?.text.trim().isNotEmpty) ??
-        false;
-    final entranceArrivals = additionalCharacters.where(hasEntrance).toList();
-    final silentArrivals = additionalCharacters
-        .where((c) => !hasEntrance(c))
-        .toList();
-    final orderedArrivals = [...entranceArrivals, ...silentArrivals];
-
-    // Decoupled model: ensure members exist for the original 1:1 character
-    // and every additional character. Without this, the group loads empty
-    // (setActiveGroup / GroupTurnManager will have no one to speak).
-    // Ported from the fix originally contributed in PR #44 by @MisterLotto.
-    await _createGroupMember(group.id, _activeCharacter!);
-    for (final c in orderedArrivals) {
-      await _createGroupMember(group.id, c);
-    }
-
-    // Create a new session for the group and copy all messages
-    final newSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-    final copiedMessages = <MessagesCompanion>[];
-    for (int i = 0; i < _messages.length; i++) {
-      final m = _messages[i];
-      // Back-fill characterId on AI messages (null in 1:1 mode)
-      String? charId = m.characterId;
-      if (!m.isUser && charId == null) {
-        charId = originalCharId;
-      }
-      copiedMessages.add(
-        MessagesCompanion(
-          sessionId: drift.Value(newSessionId),
-          position: drift.Value(i),
-          sender: drift.Value(m.sender),
-          isUser: drift.Value(m.isUser),
-          characterId: drift.Value(charId),
-          swipes: drift.Value(jsonEncode(m.swipes)),
-          swipeIndex: drift.Value(m.swipeIndex),
-          swipeDurations: drift.Value(jsonEncode(m.swipeDurations)),
-        ),
-      );
-    }
-
-    // Insert the new session
-    await _db.upsertSession(
-      SessionsCompanion.insert(
-        id: newSessionId,
-        groupId: drift.Value(group.id),
-        name: drift.Value(_sessionName),
-        description: drift.Value(_sessionDescription),
-        authorNote: drift.Value(_authorNote),
-        authorNoteDepth: drift.Value(_authorNoteStrength),
-        summary: drift.Value(_summary.isEmpty ? null : _summary),
-        summaryLastIndex: drift.Value(
-          _summaryLastIndex > 0 ? _summaryLastIndex : null,
-        ),
-        parentSession: drift.Value(_currentSessionId),
-        forkIndex: drift.Value(_messages.length - 1),
-        trustLevel: drift.Value(_relationshipService.trustLevel),
-        activeFixation: drift.Value(_relationshipService.activeFixation),
-        fixationLifespan: drift.Value(_relationshipService.fixationLifespan),
-        spatialStance: drift.Value(_relationshipService.spatialStance),
-        startDayOfWeek: drift.Value(_timeService.startDayOfWeekAnchor),
-        createdAt: drift.Value(DateTime.now()),
-        updatedAt: drift.Value(DateTime.now()),
-      ),
-    );
-    if (copiedMessages.isNotEmpty) {
-      await _db.insertMessages(copiedMessages);
-    }
-
-    debugPrint(
-      '[ChatService] \u{1F500} Forked 1:1 chat to group "${group.name}" '
-      '(${_messages.length} messages copied)',
-    );
-
-    // Switch to the new group (this loads the session we just created)
-    await setActiveGroup(group, groupRepo: groupRepo);
-
-    // Run any custom entrances WITHOUT blocking the caller, so the wizard can
-    // navigate to the group immediately and the entrance messages stream into
-    // the now-visible chat instead of waiting behind a spinner. Entrants cut in
-    // one-by-one in the order they were added.
-    if (entranceArrivals.isNotEmpty) {
-      _entrancesInFlight = true; // block user turns until the sequence finishes
-      unawaited(() async {
-        try {
-          for (final addCard in entranceArrivals) {
-            final entry = entrances[_getCharacterIdFromCard(addCard)]!;
-            final text = entry.text.trim();
-
-            // Members are copied under fresh UUIDs on fork, so resolve by name
-            // (stable) and use the resolved member's id — else the entrance
-            // attributes to the wrong member.
-            final resolved = _groupCharacters.firstWhere(
-              (c) => c.name == addCard.name,
-              orElse: () => addCard,
-            );
-            final resolvedId = _getCharacterIdFromCard(resolved);
-
-            if (entry.creative) {
-              // Hidden one-shot directive; the member writes their own entrance.
-              // Shared with live /join --full + sidebar adds via the helper.
-              final ok = await _generateMemberEntrance(resolved, text);
-              if (!ok) {
-                // Surface the failure so the user isn't left wondering why the
-                // group loaded with no entrance.
-                _messages.add(
-                  ChatMessage(
-                    text:
-                        '⚠ ${resolved.name}\'s entrance could not be generated.',
-                    sender: 'System',
-                    isUser: false,
-                  ),
-                );
-                await _saveChat();
-                notifyListeners();
-              }
-            } else {
-              // Opening line: the entrance IS the user's text, verbatim — it
-              // becomes the character's message as-is, no LLM generation.
-              _messages.add(
-                ChatMessage(
-                  text: text,
-                  sender: resolved.name,
-                  isUser: false,
-                  characterId: resolvedId,
-                ),
-              );
-              await _saveChat();
-              notifyListeners();
-            }
-          }
-        } catch (e) {
-          debugPrint('[Fork:Entrance] sequence failed: $e');
-        } finally {
-          // The entrances are one-off cut-ins. In round-robin the next turn goes
-          // to whoever falls right after the LAST entrant in the rotation order
-          // (originals, then entrance arrivals, then silent arrivals) — i.e. the
-          // first silent arrival, or wrapping back to the original if there are
-          // none. advanceAfterRegeneration parks the pointer at last-entrant + 1.
-          // Done in `finally` so a generation hiccup can't leave the rotation
-          // stuck on the entrant. Random needs no fix-up.
-          final lastEntrantName = entranceArrivals.last.name;
-          if (turnOrder == TurnOrder.roundRobin &&
-              _groupCharacters.any((c) => c.name == lastEntrantName)) {
-            final lastEntrant = _groupCharacters.firstWhere(
-              (c) => c.name == lastEntrantName,
-            );
-            _groupManager?.advanceAfterRegeneration(lastEntrant);
-          }
-          _entrancesInFlight = false; // user turns allowed again
-          // The turn pointer changed after generation finished. GroupTurnManager
-          // notifies its own listeners, but the chat UI watches ChatService, so
-          // we must propagate here — otherwise the next-speaker indicator keeps
-          // showing the (stale) entrant even though the pointer is correct.
-          notifyListeners();
-        }
-      }());
-    }
-
-    return group;
-  }
-
-  /// Create a group member (decoupled model): copy the character's avatar into
-  /// the group's private storage and insert a group_members row.
-  /// Shared by [addCharacterToGroup] (live add) and [forkToGroupChat] (initial
-  /// membership when forking a 1:1 chat into a group).
-  ///
-  /// This is the core of the fix originally contributed in PR #44 by @MisterLotto.
-  Future<void> _createGroupMember(
-    String groupId,
-    CharacterCard character,
-  ) async {
-    final mid = const Uuid().v4();
-    final avDir = Directory(
-      path.join(_storageService.groupsDir.path, groupId, 'avatars'),
-    );
-    await avDir.create(recursive: true);
-
-    await _characterRepository!.duplicateCharacter(
-      character,
-      targetDirOverride: avDir.path,
-      forcedBasename: mid,
-      skipLibraryInsert: true,
-    );
-
-    final db = await AppDatabase.instance();
-    await db.insertGroupMember(
-      GroupMembersCompanion(
-        id: drift.Value(mid),
-        groupId: drift.Value(groupId),
-        name: drift.Value(character.name),
-        description: drift.Value(character.description),
-        personality: drift.Value(character.personality),
-        scenario: drift.Value(character.scenario),
-        firstMessage: drift.Value(character.firstMessage),
-        mesExample: drift.Value(character.mesExample),
-        systemPrompt: drift.Value(character.systemPrompt),
-        postHistoryInstructions: drift.Value(character.postHistoryInstructions),
-        alternateGreetings: drift.Value(
-          jsonEncode(character.alternateGreetings),
-        ),
-        tags: drift.Value(jsonEncode(character.tags)),
-        avatarFilename: drift.Value('$mid.png'),
-        ttsVoice: drift.Value(character.ttsVoice),
-        lorebook: drift.Value(
-          character.lorebook != null
-              ? jsonEncode(character.lorebook!.toJson())
-              : null,
-        ),
-        worldNames: drift.Value(jsonEncode(character.worldNames)),
-        frontPorchExtensions: drift.Value(
-          character.frontPorchExtensions != null
-              ? jsonEncode(character.frontPorchExtensions!.toJson())
-              : null,
-        ),
-        rawExtensions: drift.Value(
-          character.rawExtensions != null
-              ? jsonEncode(character.rawExtensions!)
-              : null,
-        ),
-        memberState: drift.Value('{}'),
-      ),
-    );
-  }
-
-  /// Add a character to the currently active group chat.
-  Future<bool> addCharacterToGroup(
-    CharacterCard character,
-    GroupChatRepository groupRepo,
-  ) async {
-    if (_activeGroup == null || _characterRepository == null) return false;
-    if (_isGenerating) return false;
-
-    // Decoupled model: copy the character's avatar into the group's private
-    // storage and insert a group_members row. Shared with forkToGroupChat
-    // via _createGroupMember (ported from the fix originally contributed in
-    // PR #44 by @MisterLotto).
-    await _createGroupMember(_activeGroup!.id, character);
-
-    await groupRepo.save(_activeGroup!);
-
-    // Re-resolve from private members (decoupled).
-    final resolved = <CharacterCard>[];
-    final memberRows = await groupRepo.getMembersForGroup(_activeGroup!.id);
-    for (final m in memberRows) {
-      if (m.avatarFilename != null) {
-        final p = path.join(
-          _storageService.groupsDir.path,
-          _activeGroup!.id,
-          'avatars',
-          m.avatarFilename!,
-        );
-        if (await File(p).exists()) {
-          resolved.add(m.toCharacterCard(resolvedImagePath: p));
-        }
-      }
-    }
-    _groupManager?.refreshCharacters(resolved);
-
-    debugPrint(
-      '[ChatService] \u{2795} Added ${character.name} to group ${_activeGroup!.name}',
-    );
-    notifyListeners();
-    return true;
-  }
-
-  /// Remove a character from the currently active group chat.
-  /// Returns false if the group would have fewer than 2 characters.
-  Future<bool> removeCharacterFromGroup(
-    CharacterCard character,
-    GroupChatRepository groupRepo,
-  ) async {
-    if (_activeGroup == null || _characterRepository == null) return false;
-    if (_isGenerating) return false;
-    // Minimum member count now based on loaded group members (decoupled).
-    // (enforcement wired via member list length in calling paths)
-
-    // remove ID path removed (decoupled). Real remove deletes the GroupMember row + avatar file.
-    final charId = _getCharacterIdFromCard(character);
-    // (old removeCharacterId deleted)
-    await groupRepo.save(_activeGroup!);
-
-    // Drop any per-char state for the removed member (realism + author notes + per-char system prompts + rag priority)
-    _groupRealism.remove(charId);
-    _groupAuthorNotes.remove(charId);
-    _groupAuthorNoteStrengths.remove(charId);
-    _groupCharacterSystemPrompts.remove(charId);
-    _groupCharacterRAGPriorities.remove(charId);
-    if (_activeGroup != null) {
-      // (old checkpoint call removed in v30)
-    }
-
-    // Re-resolve after remove — old IDs path removed (decoupled).
-    final resolved = <CharacterCard>[]; // from group members
-    _groupManager?.refreshCharacters(resolved);
-
-    debugPrint(
-      '[ChatService] \u{2796} Removed ${character.name} from group ${_activeGroup!.name}',
-    );
-    notifyListeners();
-    return true;
-  }
 
   /// Returns a stable ID string for a character card.
   /// Delegates to the canonical stable ID for group contexts.
@@ -5475,680 +5128,6 @@ class ChatService extends ChangeNotifier {
     await _generateResponse(GenerationMode.normal);
   }
 
-  Future<bool> manualReprocessNeeds(int index, String critique) async {
-    if (index < 0 || index >= _messages.length) return false;
-    if (_isGenerating) return false;
-
-    final msg = _messages[index];
-    if (msg.isUser || msg.sender == 'System') return false;
-
-    final meta = msg.activeMetadata;
-    if (meta == null || !meta.containsKey('realism_state')) return false;
-
-    final preState = meta['realism_state'];
-    if (preState is! Map || preState['needs'] == null) return false;
-
-    // A: entry guard (usable needs data) already passed; button in UI also checks now.
-
-    final oldNeedsDeltas = <String, int>{};
-    Map<String, dynamic>? originalNeedsDeltasForStash;
-    if (meta.containsKey('needs_deltas')) {
-      final oldMap = meta['needs_deltas'] as Map;
-      originalNeedsDeltasForStash = Map<String, dynamic>.from(oldMap);
-      for (final k in oldMap.keys) {
-        if (oldMap[k] is Map && oldMap[k]['delta'] is num) {
-          oldNeedsDeltas[k.toString()] = (oldMap[k]['delta'] as num).toInt();
-        }
-      }
-    }
-
-    // D: stash original deltas under pre_reprocess key (before any commit)
-    final hasPrior = originalNeedsDeltasForStash != null;
-
-    final bool isLast = index == _messages.length - 1;
-    final bool isGroupNonObs = _activeGroup != null && !_observerMode;
-    String? sid;
-    CharacterCard? targetSpeakerCard;
-    if (isGroupNonObs) {
-      if (_groupCharacters.isEmpty) return false;
-      targetSpeakerCard = _groupCharacters.firstWhere(
-        (c) => c.name == msg.sender,
-        orElse: () => _groupCharacters.first,
-      );
-      sid = _getCharacterIdFromCard(targetSpeakerCard);
-    }
-
-    // Snapshot the *active* live state *before* any historical/target prepare (for strict historical meta-only)
-    final Map<String, int> preOpLive;
-    String? preOpActiveSid;
-    CharacterCard? preActiveChar;
-    if (isGroupNonObs) {
-      preOpActiveSid = _getCurrentSpeakerIdForRealism();
-      preOpLive = Map<String, int>.from(_getGroupNeeds(preOpActiveSid));
-      preActiveChar = _activeCharacter;
-    } else {
-      preOpLive = Map<String, int>.from(_needsSimulation.vector);
-    }
-
-    // Snapshot for rollback (target's at click time; used for last + compute baseline)
-    final Map<String, int> livePreClick;
-    if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-      livePreClick = Map<String, int>.from(_getGroupNeeds(sid));
-    } else {
-      livePreClick = Map<String, int>.from(_needsSimulation.vector);
-    }
-
-    // Group impersonation dance + load for prompt fidelity (personality/stance via getActiveCharacter in evaluate) + prepare scalar to msg pre.
-    // For !isLast we still dance (to get correct prompt for that speaker's critique) but will strictly rollback live after.
-    if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-      if (targetSpeakerCard != null) {
-        _activeCharacter = targetSpeakerCard;
-      }
-      _loadGroupRealismIntoScalars(sid);
-    }
-    _needsSimulation.restoreFromSnapshot(preState['needs']);
-    final Map<String, int> restoredPreVector = Map<String, int>.from(
-      _needsSimulation.vector,
-    );
-
-    if (isLast) {
-      _isVerifyingRealism = true;
-      _verificationPass = 1;
-      _verificationMaxPasses = 1;
-      notifyListeners();
-    }
-
-    _pendingRealismMetadata = null;
-    _realismEvalCancelled = false;
-    final koboldForReprocess = _llmProvider?.koboldService;
-    if (koboldForReprocess != null) {
-      await koboldForReprocess.waitForIdle();
-    }
-
-    final bool reprocessOk = await _needsImpactEvaluator
-        .reprocessWithUserCritique(msg.displayText, oldNeedsDeltas, critique);
-
-    if (!reprocessOk) {
-      // A: non-destructive: rollback to pre-op active live (historical leaves current active untouched)
-      if (isGroupNonObs &&
-          preOpActiveSid != null &&
-          preOpActiveSid.isNotEmpty) {
-        _loadGroupRealismIntoScalars(preOpActiveSid);
-      } else if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-        _setGroupNeeds(sid, livePreClick); // fallback
-      } else {
-        _needsSimulation.restoreFromSnapshot({'vector': preOpLive});
-      }
-      if (preActiveChar != null) {
-        _activeCharacter = preActiveChar;
-      }
-      if (isLast) {
-        _isVerifyingRealism = false;
-        _pendingRealismMetadata = null;
-        await _saveChat();
-        notifyListeners();
-      } else {
-        _pendingRealismMetadata = null;
-      }
-      return false;
-    }
-
-    // Success path: commit metadata always (for the target msg)
-    final updatedMeta = Map<String, dynamic>.from(meta);
-    final computed = _needsSimulation.computeNeedsDeltasWithReasons(
-      restoredPreVector,
-    );
-    updatedMeta['needs_deltas'] = computed;
-
-    // Keep realism_state['needs'] aligned with manual corrections so swipe
-    // navigation and regen do not resurrect a stale pre-reprocess vector.
-    if (_needsSimEnabled && updatedMeta['realism_state'] is Map) {
-      final postReprocessVector = isGroupNonObs && sid != null && sid.isNotEmpty
-          ? Map<String, int>.from(_getGroupNeeds(sid))
-          : Map<String, int>.from(_needsSimulation.vector);
-      final rs = Map<String, dynamic>.from(updatedMeta['realism_state'] as Map);
-      final needsSnap = <String, dynamic>{'vector': postReprocessVector};
-      if (computed.isNotEmpty) {
-        needsSnap['deltas'] = computed;
-      }
-      rs['needs'] = needsSnap;
-      updatedMeta['realism_state'] = rs;
-    }
-
-    if (hasPrior) {
-      updatedMeta['needs_deltas_pre_reprocess'] = originalNeedsDeltasForStash;
-    }
-    // D: record critique in verif meta (for history/tooltip)
-    final verifMeta =
-        (updatedMeta[RealismVerification.kMetaKey] as Map<String, dynamic>?) ??
-        <String, dynamic>{};
-    verifMeta['reprocess_critique'] = critique;
-    updatedMeta[RealismVerification.kMetaKey] = verifMeta;
-
-    if (_pendingRealismMetadata != null) {
-      for (final e in _pendingRealismMetadata!.entries) {
-        updatedMeta[e.key] = e.value;
-      }
-    }
-
-    msg.swipeMetadata[msg.swipeIndex] = updatedMeta;
-
-    if (isLast) {
-      // Persist live only for current last speaker
-      if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-        _saveScalarsIntoGroupRealism(sid);
-      }
-      _isVerifyingRealism = false;
-      _pendingRealismMetadata = null;
-    } else {
-      // Historical: meta updated; strictly no live/group mutation left behind.
-      // Transient onSaveChat may fire from temp apply (critique path); scalars rolled back; _saveChat here is only for target msg metadata.
-      if (isGroupNonObs &&
-          preOpActiveSid != null &&
-          preOpActiveSid.isNotEmpty) {
-        _loadGroupRealismIntoScalars(preOpActiveSid);
-      } else if (!isGroupNonObs) {
-        _needsSimulation.restoreFromSnapshot({'vector': preOpLive});
-      }
-      if (preActiveChar != null) {
-        _activeCharacter = preActiveChar;
-      }
-      _pendingRealismMetadata = null;
-    }
-    await _saveChat();
-    notifyListeners();
-    return true;
-  }
-
-  /// Revert a prior manual reprocess on the given message.
-  /// Restores the stashed 'needs_deltas_pre_reprocess' into 'needs_deltas' (with reasons),
-  /// rewinds live scalars *only if last* (historical: metadata-only update, live scalars/group untouched).
-  /// Safe for 1:1 and group (speaker via msg.sender + dance on last).
-  /// Returns true on success.
-  Future<bool> revertNeedsReprocess(int index) async {
-    if (index < 0 || index >= _messages.length) return false;
-    if (_isGenerating) return false;
-    final msg = _messages[index];
-    final meta = msg.activeMetadata;
-    if (meta == null || !meta.containsKey('needs_deltas_pre_reprocess')) {
-      return false;
-    }
-    final preState = meta['realism_state'];
-    if (preState is! Map || preState['needs'] == null) return false;
-
-    final stashedDeltasMap = meta['needs_deltas_pre_reprocess'] as Map;
-    final oldDeltas = <String, int>{};
-    stashedDeltasMap.forEach((k, v) {
-      if (v is Map && v['delta'] is num) {
-        oldDeltas[k.toString()] = (v['delta'] as num).toInt();
-      }
-    });
-
-    final bool isLast = index == _messages.length - 1;
-    final bool isGroupNonObs = _activeGroup != null && !_observerMode;
-    String? sid;
-    CharacterCard? targetSpeakerCard;
-    if (isGroupNonObs) {
-      if (_groupCharacters.isEmpty) return false;
-      targetSpeakerCard = _groupCharacters.firstWhere(
-        (c) => c.name == msg.sender,
-        orElse: () => _groupCharacters.first,
-      );
-      sid = _getCharacterIdFromCard(targetSpeakerCard);
-    }
-
-    // Snapshot pre-op active for strict !isLast rollback (no live mutation)
-    CharacterCard? preActiveChar;
-    Map<String, int> preOpLive = {};
-    String? preOpSid;
-    if (isGroupNonObs) {
-      preOpSid = _getCurrentSpeakerIdForRealism();
-      preOpLive = Map<String, int>.from(_getGroupNeeds(preOpSid));
-      preActiveChar = _activeCharacter;
-    } else {
-      preOpLive = Map<String, int>.from(_needsSimulation.vector);
-    }
-
-    // Dance + prepare only if we will keep the effect (isLast); for historical we temp dance but rollback
-    if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-      if (targetSpeakerCard != null) _activeCharacter = targetSpeakerCard;
-      _loadGroupRealismIntoScalars(sid);
-    }
-    _needsSimulation.restoreFromSnapshot(preState['needs'] as Map);
-
-    if (oldDeltas.isNotEmpty) {
-      String? reasonToRestore;
-      final values = (stashedDeltasMap as Map?)?.values ?? const [];
-      for (final v in values) {
-        if (v is Map &&
-            v['reason'] is String &&
-            (v['reason'] as String).isNotEmpty) {
-          reasonToRestore = v['reason'] as String;
-          break;
-        }
-      }
-      _needsSimulation.applySceneImpact(
-        NeedsImpact(deltas: oldDeltas, reason: reasonToRestore),
-      );
-    }
-
-    final updated = Map<String, dynamic>.from(meta);
-    updated['needs_deltas'] = stashedDeltasMap;
-    updated.remove('needs_deltas_pre_reprocess');
-    if (_needsSimEnabled && updated['realism_state'] is Map) {
-      final postRevertVector = isGroupNonObs && sid != null && sid.isNotEmpty
-          ? Map<String, int>.from(_getGroupNeeds(sid))
-          : Map<String, int>.from(_needsSimulation.vector);
-      final rs = Map<String, dynamic>.from(updated['realism_state'] as Map);
-      final needsSnap = <String, dynamic>{'vector': postRevertVector};
-      if (stashedDeltasMap.isNotEmpty) {
-        needsSnap['deltas'] = stashedDeltasMap;
-      }
-      rs['needs'] = needsSnap;
-      updated['realism_state'] = rs;
-    }
-    msg.swipeMetadata[msg.swipeIndex] = updated;
-
-    if (isLast) {
-      if (isGroupNonObs && sid != null && sid.isNotEmpty) {
-        _saveScalarsIntoGroupRealism(sid);
-      }
-    } else {
-      // Historical meta-only: rollback live + char, do not persist to group for this op
-      if (isGroupNonObs && preOpSid != null && preOpSid.isNotEmpty) {
-        _loadGroupRealismIntoScalars(preOpSid);
-      } else if (!isGroupNonObs) {
-        _needsSimulation.restoreFromSnapshot({'vector': preOpLive});
-      }
-      if (preActiveChar != null) _activeCharacter = preActiveChar;
-    }
-
-    await _saveChat();
-    notifyListeners();
-    return true;
-  }
-
-  Future<void> regenerateLastMessage() async {
-    if (_messages.isEmpty || _isGenerating || _guestBusy) return;
-
-    // Check if the last message is from the character
-    if (!_messages.last.isUser && _messages.last.sender != 'System') {
-      // Instead of removing the message, we generate a new swipe
-      // Temporarily remove the last message so the prompt doesn't include it
-      final lastMsg = _messages.removeLast();
-      // Is this a Scene Guest message? If so the whole regen must stay a
-      // parity-safe GUEST turn: skip every Realism/Needs revert + re-eval below
-      // and regenerate spoken as the guest (guestSpeaker), exactly like the
-      // normal guest turn path. A host message keeps the existing behaviour.
-      final regenGuest = _sceneGuestForMessage(lastMsg);
-      // If the message was authored by a guest who has since LEFT the scene (or
-      // had their card deleted), we can neither regenerate as them nor run the
-      // host realism block on their text (that would perturb the host's
-      // bond/trust/emotion/needs from guest-authored content). Refuse cleanly.
-      if (regenGuest == null && _isGuestAuthoredMessage(lastMsg)) {
-        _messages.add(lastMsg); // put it back untouched
-        notifyListeners();
-        _setGuestStatus(
-          'Can’t regenerate "${lastMsg.sender}" — they have left the scene.',
-          isError: true,
-        );
-        return;
-      }
-      // Snapshot the rejected swipe's metadata (e.g. manual needs reprocess) before
-      // we add a new swipe — regen must not clobber prior swipe timelines.
-      final rejectedSwipeIndex = lastMsg.swipeIndex;
-      Map<String, dynamic>? preservedRejectedMeta;
-      if (rejectedSwipeIndex >= 0) {
-        if (rejectedSwipeIndex < lastMsg.swipeMetadata.length &&
-            lastMsg.swipeMetadata[rejectedSwipeIndex] != null) {
-          preservedRejectedMeta = Map<String, dynamic>.from(
-            lastMsg.swipeMetadata[rejectedSwipeIndex]!,
-          );
-        } else if (lastMsg.activeMetadata != null) {
-          preservedRejectedMeta = Map<String, dynamic>.from(
-            lastMsg.activeMetadata!,
-          );
-        }
-      }
-      notifyListeners();
-
-      // In group mode, force the turn manager to the *original* speaker of the
-      // removed message before generation. This prevents regen from picking a
-      // different character (the core of the "speaker changed after regen" bug).
-      if (_activeGroup != null) {
-        final originalSpeaker = _groupCharacters.firstWhere(
-          (c) => c.name == lastMsg.sender,
-          orElse: () => _groupCharacters.first,
-        );
-        _groupManager?.setNextSpeaker(originalSpeaker);
-      }
-
-      // Revert realism state from the rejected swipe and re-evaluate
-      // (host turns only — guest messages carry no Realism/Needs state).
-      if (_realismEnabled && _activeGroup == null && regenGuest == null) {
-        // CRITICAL FIX: Find the baseline realism state from the previous accepted message.
-        // We want to use the final state of the LAST ACCEPTED character message as our baseline,
-        // not just blindly revert deltas and re-evaluate from scratch.
-        Map<String, dynamic>? previousMessageState;
-        if (_messages.length >= 2) {
-          // Look back through messages to find the last bot message before the one we're regenerating
-          for (int i = _messages.length - 1; i >= 0; i--) {
-            if (!_messages[i].isUser && _messages[i].sender != 'System') {
-              final meta = _messages[i].activeMetadata;
-              if (meta != null && meta.containsKey('realism_state')) {
-                previousMessageState =
-                    meta['realism_state'] as Map<String, dynamic>;
-                debugPrint(
-                  '[Realism:Regen] Found previous accepted message baseline state at message index $i',
-                );
-                break;
-              }
-            }
-          }
-        }
-
-        bool wasNudged = false;
-        if (lastMsg.activeMetadata != null &&
-            lastMsg.activeMetadata!['realism_state'] is Map) {
-          wasNudged =
-              lastMsg.activeMetadata!['realism_state']['time_nudged'] == true;
-        }
-
-        if (lastMsg.activeMetadata != null) {
-          final bondDelta = lastMsg.activeMetadata!['bond_delta'] as int? ?? 0;
-          final moodDelta = lastMsg.activeMetadata!['mood_delta'] as int? ?? 0;
-          final arousalDelta =
-              lastMsg.activeMetadata!['arousal_delta'] as int? ?? 0;
-          final trustDelta =
-              lastMsg.activeMetadata!['trust_delta'] as int? ?? 0;
-
-          if (bondDelta != 0) {
-            _relationshipService.applyScoreDelta(-bondDelta);
-          }
-          if (moodDelta != 0) {
-            _moodDecayCounter = 0;
-          }
-          if (trustDelta != 0) {
-            _relationshipService.setTrustLevelForRevert(
-              (_relationshipService.trustLevel - trustDelta).clamp(-100, 100),
-            );
-          }
-
-          // Revert climax state if this response triggered refractory cooldown.
-          // The climax checker stores the pre-climax arousal so we can restore it.
-          final climaxTriggered =
-              lastMsg.activeMetadata!['climax_triggered'] as bool? ?? false;
-          if (climaxTriggered && _nsfwService.nsfwCooldownEnabled) {
-            final preClimaxArousal =
-                lastMsg.activeMetadata!['pre_climax_arousal'] as int? ?? 0;
-            _nsfwService.setArousalLevel(preClimaxArousal);
-            _nsfwService.setCooldownTurnsRemaining(0);
-            _nsfwService.setCooldownTurnsTotal(0);
-            debugPrint(
-              '[Realism:Regen] Reverted climax state: arousal restored to $preClimaxArousal, cooldown cleared',
-            );
-          } else if (arousalDelta != 0 && _nsfwService.nsfwCooldownEnabled) {
-            // Normal arousal delta revert (no climax involved)
-            _nsfwService.setArousalLevel(
-              (_nsfwService.arousalLevel - arousalDelta).clamp(-100, 100),
-            );
-          }
-
-          // Needs pre-turn vector revert — mirrors the bond/trust/arousal delta
-          // system so regen can undo the decay + fulfillment that ran for this
-          // user turn, even when the previous message's realism_state snapshot
-          // lacks a 'needs' entry (e.g. needs was enabled mid-chat).
-          final preTurnNeeds =
-              lastMsg.activeMetadata!['needs_pre_turn_vector'] as Map?;
-          if (preTurnNeeds != null && _needsSimEnabled) {
-            _needsSimulation.restoreFromSnapshot({
-              'vector': Map<String, int>.from(preTurnNeeds),
-            });
-            debugPrint(
-              '[Realism:Regen] Restored needs vector from pre-turn snapshot on rejected message',
-            );
-          }
-        }
-
-        // CRITICAL FIX: Restore the baseline state from the previous accepted message.
-        // This ensures the new regenerated message is evaluated against the correct baseline,
-        // not from scratch which would produce wildly different realism values.
-        if (previousMessageState != null) {
-          _relationshipService.restoreFromMessageState(previousMessageState);
-          _moodDecayCounter =
-              previousMessageState['moodDecayCounter'] as int? ??
-              _moodDecayCounter;
-          _characterEmotion =
-              previousMessageState['characterEmotion'] as String? ??
-              _characterEmotion;
-          _emotionIntensity =
-              previousMessageState['emotionIntensity'] as String? ??
-              _emotionIntensity;
-
-          _timeService.restoreTimeForSwipeOrRegen(
-            previousMessageState,
-            wasNudged: wasNudged,
-          );
-
-          _nsfwService.restoreNsfwFromMessageState(previousMessageState);
-
-          // Needs simulation snapshot (clean port)
-          // Guard + no enabled override: prevents stale resurrection on regen after toggle-off.
-          if (previousMessageState.containsKey('needs') &&
-              previousMessageState['needs'] is Map &&
-              _needsSimEnabled) {
-            final needsData = previousMessageState['needs'] as Map;
-            if (needsData['vector'] is Map) {
-              final vector = Map<String, int>.from(needsData['vector'] as Map);
-              _needsSimulation.restoreFromSnapshot({'vector': vector});
-            }
-          }
-
-          debugPrint(
-            '[Realism:Regen] ✓ Restored baseline from previous accepted message: bond=${_relationshipService.affectionScore}, emotion=$_characterEmotion, trust=${_relationshipService.trustLevel}, arousal=${_nsfwService.arousalLevel}',
-          );
-        } else {
-          debugPrint(
-            '[Realism:Regen] ⚠ No previous message baseline found, continuing with current reverted state',
-          );
-        }
-        // Set UI streaming state
-        _isEvaluatingRealism = true;
-        _realismEvalStreamText = '';
-        notifyListeners();
-
-        void handleChunk(String chunk) {
-          _realismEvalStreamText += chunk;
-          // Debounce: coalesce rapid token arrivals into one rebuild per 150 ms
-          _evalChunkTimer?.cancel();
-          _evalChunkTimer = Timer(const Duration(milliseconds: 150), () {
-            try {
-              notifyListeners();
-            } catch (_) {
-              // Widget was deactivated — timer fired after navigation
-            }
-          });
-        }
-
-        // Apply decay and cooldown — mirrors the normal path (lines 3933-3937).
-        // This ensures _needsVector differs from the saved pre-turn vector
-        // so post-generation deltas are non-zero.
-        _applyMoodDecay();
-        _needsSimulation.tickDecay();
-        _nsfwService.decrementCooldownIfActive();
-
-        // Record the (restored) needs baseline as the pre-turn vector BEFORE
-        // generation so the post-generation checks can compute proper deltas.
-        if (_needsSimEnabled && _needsSimulation.vector.isNotEmpty) {
-          _pendingRealismMetadata ??= {};
-          _pendingRealismMetadata!['needs_pre_turn_vector'] =
-              Map<String, int>.from(_needsSimulation.vector);
-        }
-
-        if (_storageService.realismSettings.realismOneShotEval) {
-          await _evaluateOneShotCall(onChunk: handleChunk);
-        } else {
-          _realismEvals.beginCollectForBatchedVerification();
-          await _fireStaggeredRealismEvals(handleChunk);
-          await _realismEvals.finalizeBatchedRealismVerifications();
-
-          // Mirror of the primary send path: apply the one director batch (or the cheap
-          // accepted-originals path) so relationship/emotional/narrative deltas and side
-          // effects (bond/trust/arousal chips, fixation, autonomous objectives, emotion)
-          // are produced for swiped/regenerated assistant messages too.
-          final collected = _realismEvals.getCollectedForBatch();
-          if (collected.isNotEmpty) {
-            final items = collected
-                .map(
-                  (p) => (
-                    evalKind: p['kind'] as String,
-                    rawOutput: p['raw'] as String,
-                    sceneResponse: p['scene'] as String,
-                    preState: null,
-                    activeChar: _activeCharacter,
-                    activeGroup: _activeGroup,
-                    recentMessages: _messages,
-                    promptText: p['prompt'] as String?,
-                    injections: (p['injections'] as Map?)
-                        ?.cast<String, String>(),
-                    strictnessOverride: null,
-                    maxPassesOverride: null,
-                  ),
-                )
-                .toList();
-            final batchRes = await _realismVerifier.verifyBatch(items);
-            await _realismEvals.applyBatchResults(batchRes);
-          }
-        }
-
-        // Check for cancellation after evals complete
-        if (_realismEvalCancelled) {
-          debugPrint(
-            '[Realism] Evaluation cancelled during regenerate, aborting',
-          );
-          _realismEvalCancelled = false;
-          _evalChunkTimer?.cancel();
-          _evalChunkTimer = null;
-          _isEvaluatingRealism = false;
-          notifyListeners();
-          return;
-        }
-
-        // Cancel any pending debounce notify before closing the overlay
-        _evalChunkTimer?.cancel();
-        _evalChunkTimer = null;
-        _isEvaluatingRealism = false;
-        notifyListeners();
-      }
-
-      // In group mode the per-speaker realism eval (and its metadata / needs deltas)
-      // happens inside _generateResponse via _evaluateRealismForUpcomingSpeaker
-      // for the correctly-forced speaker. Skip the 1:1 scalar synthesis here.
-      Map<String, int>? regenPreTurn;
-      Map<String, dynamic>? needsDeltas;
-      if (_activeGroup == null && regenGuest == null) {
-        // Save pre-turn vector BEFORE _generateResponse (which clears
-        // _pendingRealismMetadata).
-        regenPreTurn =
-            _pendingRealismMetadata?['needs_pre_turn_vector']
-                as Map<String, int>?;
-
-        // Synthesize metadata after all regen evals complete — mirrors the
-        // normal path (line 4020) so emotion_label and realism_state are in
-        // _pendingRealismMetadata before _generateResponse consumes it.
-        _pendingRealismMetadata ??= {};
-        _pendingRealismMetadata!['emotion_label'] = _characterEmotion;
-        _pendingRealismMetadata!['realism_state'] = _captureRealismState(
-          preTurn: regenPreTurn,
-        );
-
-        // If cancellation was requested during realism evaluation, abort generation
-        if (_realismEvalCancelled) {
-          _realismEvalCancelled = false;
-          notifyListeners();
-          return;
-        }
-      }
-
-      // Invalidate ONNX cache for the new response (delegated)
-      _expressionService.invalidateOnnxCacheForNewResponse();
-
-      // Generate into a new message — it will be appended by _generateResponse.
-      // For a guest message we pass guestSpeaker so the new swipe is spoken as
-      // the guest and the entire Realism/Needs post-gen block is skipped (the
-      // `guestSpeaker == null` guard). For a host message regenGuest is null and
-      // this is the unchanged host path: _generateResponse runs the post-gen
-      // needs checks (climax, sexual, daily, fulfillment) that modify the needs
-      // vector, so we compute needs_deltas AFTER generation below.
-      await _generateResponse(GenerationMode.normal, guestSpeaker: regenGuest);
-
-      // Compute needs_deltas AFTER generation so the post-generation checks
-      // are reflected. This mirrors the normal generation path (line ~4053).
-      // Apply directly to the message since _pendingRealismMetadata was consumed.
-      // (For groups, the per-speaker path inside generate already attached the
-      // correct per-character needs_deltas; we only compute scalar here for 1:1.)
-      if (_activeGroup == null &&
-          regenGuest == null &&
-          _needsSimEnabled &&
-          _needsSimulation.vector.isNotEmpty) {
-        needsDeltas = _needsSimulation.computeNeedsDeltasWithReasons(
-          regenPreTurn ?? const <String, int>{},
-        );
-      }
-
-      // After generation, merge the new response as a swipe on the original message
-      if (_messages.isNotEmpty &&
-          !_messages.last.isUser &&
-          _messages.last.sender != 'System') {
-        final newText = _messages.last.text;
-        final newMetadata = _messages.last.activeMetadata != null
-            ? Map<String, dynamic>.from(_messages.last.activeMetadata!)
-            : null;
-        _messages.removeLast();
-        lastMsg.swipes.add(newText);
-        while (lastMsg.swipeDurations.length < lastMsg.swipes.length) {
-          lastMsg.swipeDurations.add(0);
-        }
-        final newSwipeIndex = lastMsg.swipes.length - 1;
-        if (preservedRejectedMeta != null && rejectedSwipeIndex >= 0) {
-          while (lastMsg.swipeMetadata.length <= rejectedSwipeIndex) {
-            lastMsg.swipeMetadata.add(null);
-          }
-          lastMsg.swipeMetadata[rejectedSwipeIndex] = preservedRejectedMeta;
-        }
-        while (lastMsg.swipeMetadata.length <= newSwipeIndex) {
-          lastMsg.swipeMetadata.add(null);
-        }
-        lastMsg.swipeIndex = newSwipeIndex;
-        // New swipe metadata only — prior swipes (incl. manual reprocess) stay intact.
-        if (needsDeltas != null && needsDeltas.isNotEmpty) {
-          lastMsg.swipeMetadata[newSwipeIndex] = {
-            ...(newMetadata ?? {}),
-            'needs_deltas': needsDeltas,
-          };
-        } else if (newMetadata != null) {
-          lastMsg.swipeMetadata[newSwipeIndex] = newMetadata;
-        }
-        _messages.add(lastMsg);
-        // Host messages restore the active character's Realism/Needs from the
-        // accepted swipe; guest messages carry none, so leave host state intact.
-        if (regenGuest == null) _restoreRealismStateFromMessage(lastMsg);
-        await _saveChat();
-        notifyListeners();
-
-        // In group mode, advance the turn pointer past the regenerated speaker
-        // so the next natural generation continues the correct rotation instead
-        // of repeating the same character.
-        if (_activeGroup != null) {
-          final originalSpeaker = _groupCharacters.firstWhere(
-            (c) => c.name == lastMsg.sender,
-            orElse: () => _groupCharacters.first,
-          );
-          _groupManager?.advanceAfterRegeneration(originalSpeaker);
-        }
-      }
-    }
-  }
 
   /// Navigate swipes on a specific message. direction: -1 = left, +1 = right.
   /// If swiping right past the last swipe on the last bot message, regenerates.
